requirements.txt
anyio==4.13.0
APScheduler==3.11.2
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Flask==3.1.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.2.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
python-telegram-bot==22.7
reportlab==4.4.10
SQLAlchemy==2.0.49
typing_extensions==4.15.0
tzlocal==5.3.1
Werkzeug==3.1.8
